Fix crashes in tests of random numbers
I identified and fixed two issues. Hopefully, these were
the real underlying issues of the crashes.

Firstly, there was a signed integer overflow which is an undefined
behaviour. I fixed it by using different computation for signed
and unsigned types.

Secondly, the function demangle was returning a pointer to a local buffer
and also using __cxa_demangle the wrong way (it expects a malloced pointer
not a pointer to stack).